+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- **CUMULUS-1895**
+  - Update execution events to display more details in modal
+
+- **CUMULUS-2169**
+  - Add provider drop down to granule tables to filter by provider
+
+- **CUMULUS-2206**
+  - Implement lazy loading for SortableTable, Datepicker, and Sidebar
+  - Improve performance
+
+### Changed
+
+- **CUMULUS-2251**
+  - Add the `Granule Actions` button
+
 ## [v3.0.0]
 
 ### Fixed